requirements.txt
pycord
discord.py
requests